Move pvc clone specific functions to pvc-clone-controller
Now that the validation is no longer done in the webhook
no need for that code to be in the controller common util
file. Moved the UT accordingly.
