+++ pkg/controller/common/util.go
@@ -966,50 +966,6 @@ func validateTokenData(tokenData *token.Payload, srcNamespace, srcName, targetNa
 	return nil
 }
 
-// validateContentTypes compares the content type of a clone DV against its source PVC's one
-func validateContentTypes(sourcePVC *corev1.PersistentVolumeClaim, spec *cdiv1.DataVolumeSpec) (bool, cdiv1.DataVolumeContentType, cdiv1.DataVolumeContentType) {
-	sourceContentType := GetPVCContentType(sourcePVC)
-	targetContentType := spec.ContentType
-	if targetContentType == "" {
-		targetContentType = cdiv1.DataVolumeKubeVirt
-	}
-	return sourceContentType == targetContentType, sourceContentType, targetContentType
-}
-
-// ValidateClone compares a clone spec against its source PVC to validate its creation
-func ValidateClone(sourcePVC *corev1.PersistentVolumeClaim, spec *cdiv1.DataVolumeSpec) error {
-	var targetResources corev1.VolumeResourceRequirements
-
-	valid, sourceContentType, targetContentType := validateContentTypes(sourcePVC, spec)
-	if !valid {
-		msg := fmt.Sprintf("Source contentType (%s) and target contentType (%s) do not match", sourceContentType, targetContentType)
-		return errors.New(msg)
-	}
-
-	isSizelessClone := false
-	explicitPvcRequest := spec.PVC != nil
-	if explicitPvcRequest {
-		targetResources = spec.PVC.Resources
-	} else {
-		targetResources = spec.Storage.Resources
-		// The storage size in the target DV can be empty
-		// when cloning using the 'Storage' API
-		if _, ok := targetResources.Requests[corev1.ResourceStorage]; !ok {
-			isSizelessClone = true
-		}
-	}
-
-	// TODO: Spec.Storage API needs a better more complex check to validate clone size - to account for fsOverhead
-	// simple size comparison will not work here
-	if (!isSizelessClone && GetVolumeMode(sourcePVC) == corev1.PersistentVolumeBlock) || explicitPvcRequest {
-		if err := ValidateRequestedCloneSize(sourcePVC.Spec.Resources, targetResources); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
 // ValidateSnapshotClone compares a snapshot clone spec against its source snapshot to validate its creation
 func ValidateSnapshotClone(sourceSnapshot *snapshotv1.VolumeSnapshot, spec *cdiv1.DataVolumeSpec) error {
 	var sourceResources, targetResources corev1.VolumeResourceRequirements
